third_party/blink/web_tests/external/wpt/css/css-animations/parsing/keyframe-selectors-expected.txt

This is a testharness.js-based test.
[FAIL] 'calc(10%)' should be a valid keyframe selector
  assert_true: expected true got false
[FAIL] 'calc(10% * sibling-index())' should be a valid keyframe selector
  assert_true: expected true got false
[FAIL] 'calc(10% * sign(1em - 1px))' should be a valid keyframe selector
  assert_true: expected true got false
[FAIL] 'entry calc(10%)' should be a valid keyframe selector
  assert_true: expected true got false
[FAIL] 'exit calc(60%)' should be a valid keyframe selector
  assert_true: expected true got false
Harness: the test ran to completion.
